[FTR] - Finalizando e corrigindo crud de propriedades rurais

+++ app/Services/Producer/DeleteProducerService.php
@@ -8,6 +8,10 @@ class DeleteProducerService
 {
     public function run(Producer $producer): bool
     {
+        if ($producer->ruralProperty->count()) {
+            throw new \Exception("Não é possivel deletar o produtor");
+        }
+
         return $producer->delete();
     }
 }
