+++ app/Services/RuralProperty/DeleteRuralPropertyService.php
@@ -8,6 +8,11 @@ class DeleteRuralPropertyService
 {
     public function run(RuralProperty $ruralProperty): bool
     {
+
+        if( $ruralProperty->herds()->count() || $ruralProperty->productionUnits()->count()) {
+            throw new \Exception("Não foi possivel deletar a propriedade");
+        }
+
         return $ruralProperty->delete();
     }
 }
